Move the distance module to the utils folder
